requirements.txt
google-api-python-client>=2.136.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
pytest
msal>=1.30.0
requests>=2.32.3